requirements.txt
# Core Dependencies
torch>=2.0.0
torch3d>=0.7.0
numpy>=1.21.0
scipy>=1.7.0
pillow>=9.0.0

# 3D and Graphics
kaolin>=0.13.0
open3d>=0.15.0
three.js>=0.150.0
pybullet>=3.2.5

# AI and Machine Learning
transformers>=4.30.0
jax>=0.4.0
jaxlib>=0.4.0
stable-diffusion-pytorch>=2.0.0
clip @ git+https://github.com/openai/CLIP.git
whisper>=1.0.0
wav2vec2>=0.1.0

# Physics and Simulation
physx>=5.1.0
ray[rllib]>=2.5.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Testing and Development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0 